backend/requirements.txt
flask
requests
videosdk
PyJWT
flask-sqlalchemy
psycopg2-binary
flask-migrate
flask-cors
python-dotenv
werkzeug
